[CHORE] mapper 확장 함수 이동
- SelectedUser <-> User

Resolves: TICO-310
Related to: TICO-311
