+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/model/SelectedUser.kt
@@ -7,3 +7,15 @@ data class SelectedUser(
     val profileUrl: String? = null,
     val selected: Boolean
 )
+
+fun User.toSelectedUser(selected: Boolean = false): SelectedUser =
+    SelectedUser(
+        id = this.id,
+        name = this.name,
+        email = email,
+        profileUrl = this.profileUrl,
+        selected = selected
+    )
+
+fun SelectedUser.toUser(): User =
+    User(id = this.id, name = this.name, email = email, profileUrl = this.profileUrl)
